Running TensorFlow model per image

+++ src/server/segmentation/models.py
@@ -31,7 +31,7 @@ def image_post_delete(sender, instance, **kwargs):
     if instance.file_obj:
         if os.path.isfile(instance.file_obj.path):
             os.remove(instance.file_obj.path)
-        if os.path.isfile(instance.pred_file_path):
+        if instance.pred_file_path and os.path.isfile(instance.pred_file_path):
             os.remove(instance.pred_file_path)
